requirements.txt
streamlit
openai
PyMuPDF
python-docx
scikit-learn
numpy
